+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/repository/Paciente_comorbidadeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.edu.unoesc.pandemicstats.springboot.model.Paciente_comorbidade;
 
 @Repository
-public interface Paciente_comorbidadeRepository extends JpaRepository<Paciente_comorbidade, Integer> {
+public interface Paciente_comorbidadeRepository extends JpaRepository<Paciente_comorbidade, Long> {
 
 }
