+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/repository/UsuarioRepository.java
@@ -12,20 +12,20 @@ import org.springframework.stereotype.Repository;
 import br.edu.unoesc.pandemicstats.springboot.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
 	@Query(value = "select u from Usuario u where u.cpfusu = ?1")
-	Usuario findByCPF(int cpf);
+	Usuario findByCPF(long cpf);
 
 	@Query(value = "select u from Usuario u where u.emausu = ?1")
 	Usuario findByEmail(String email);
 	
 	@Query(value = "select u from Usuario u where cnpjemp.cnpjemp = ?1")
-	List<Usuario> findByCnpjEmp(int cnpj);
+	List<Usuario> findByCnpjEmp(long cnpj);
 	
 	@Transactional
 	@Modifying(clearAutomatically = true)
 	@Query(value = "update Usuario set cnpjemp = null where cnpjemp.cnpjemp = ?1")
-	void setCnpjempNull(int cnpj);
+	void setCnpjempNull(long cnpj);
 	
 }
